app/botrunner/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
websockets>=12.0
mss>=9.0.0
imagehash>=4.3.0
Pillow>=10.4.0
easyocr>=1.7.0
ultralytics>=8.2.0
pyyaml>=6.0
numpy
pyautogui>=0.9.54
Appium-Python-Client>=4.0.0
